File separation
• main.py
- constants and profile related stuff moved to proper files
- implemented run method on the Application class
- other changes to support the new imports

• constants.py
- all constants moved to here

• profile_handler.py
- profile related class and functions moved to here

• user_settings.txt
- must create a new user settings to support the new file separation
