+++ src/constants.py
@@ -1,0 +1,32 @@
+import os
+
+
+PROFILE_PATH = os.path.join(os.getcwd(), 'profiles/')
+
+REGIONS = {
+    'na': ['US West (Oregon)',
+           'US West (N. California)',
+           'US East (N. Virginia)',
+           'US Central (Texas)',
+           'US Central (Illinois)',
+           'US Central (Georgia)'],
+    'eu': ['Frankfurt',
+           'Paris',
+           'Stockholm',
+           'Istanbul',
+           'London',
+           'Tokyo',
+           'Warsaw',
+           'Madrid',
+           'Bahrain'],
+    'latam': ['Santiago',
+              'Mexico City',
+              'Miami'],
+    'br': ['Sao Paulo'],
+    'ap': ['Hong Kong',
+           'Tokyo',
+           'Singapore',
+           'Sydney',
+           'Mumbai'],
+    'kr': ['Seoul']
+}
